Added more computer logic
Currently if a computer flips over a jack, then that computer is the one that slaps. If the user flips over a jack, then the next computer in the rotation does the slap. The computer slaps and adds the cards to their hand successfully.

Unfortunately, the computer slaps right as the Jack comes up. I only had time to try a couple of things, but whatever I tried to make the computer wait also froze up the entire program for the duration, so I was unable to click the Slap button while also trying to make the computer wait. I know that here would be a good place for threads, but I'm not very familiar with how threading works (I've only done stuff with it very minorly in CS 3100) so I didn't know how to implement it.